Logging: Reorganise module
