+++ servicelayer/util.py
@@ -6,7 +6,6 @@ import json
 from banal import ensure_list
 from banal.dicts import clean_dict
 
-from servicelayer import settings
 
 QUEUE_EXPIRE = 84600 * 14
 log = logging.getLogger(__name__)
@@ -82,12 +81,3 @@ def load_json(encoded):
     if encoded is None or encoded == '':
         return
     return json.loads(encoded)
-
-
-def configure_logging():
-    """setup logging configuration and return current log level"""
-    level = getattr(logging, settings.LOGGING_LEVEL, None)
-    if level is None:
-        raise ValueError('Invalid log level: %s' % settings.LOGGING_LEVEL)
-    logging.basicConfig(level=level, format=settings.LOGGING_FORMAT)
-    return level
